feat(modules/kubernetes): Monitor kubeconfig files changes
When working on a shared environment, the kubeconfig file may change due to 'kubectl config' or kubectx commands.
This may get one to execute action on a different target.

With the modification-time monitoring of the files, we can detect and prevent commands to be run when a change is detected